chore: lowercase email in emails with display name
